backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
tensorflow==2.17.0
numpy==1.26.4
opencv-python==4.10.0.84
pillow==10.4.0
pydantic==2.9.2